+++ src/lib.rs
@@ -43,7 +43,9 @@ impl From<cxx::ScriptError> for Error {
     #[allow(non_upper_case_globals)]
     fn from(err_code: cxx::ScriptError) -> Self {
         match err_code {
-            cxx::ScriptError_t_SCRIPT_ERR_UNKNOWN_ERROR => Error::Ok(ScriptError::UnknownError),
+            cxx::ScriptError_t_SCRIPT_ERR_UNKNOWN_ERROR => {
+                Error::Ok(zcash_script::AMBIGUOUS_UNKNOWN_NUM_HIGHS_ERROR)
+            }
             cxx::ScriptError_t_SCRIPT_ERR_EVAL_FALSE => Error::Ok(ScriptError::EvalFalse),
             cxx::ScriptError_t_SCRIPT_ERR_OP_RETURN => Error::Ok(ScriptError::OpReturn),
 
@@ -226,26 +228,10 @@ pub fn check_verify_callback<T: ZcashScript, U: ZcashScript>(
     )
 }
 
-fn normalize_script_error(err: ScriptError) -> ScriptError {
-    match err {
-        ScriptError::BadOpcode(Some(_)) => ScriptError::BadOpcode(None),
-        ScriptError::DisabledOpcode(Some(_)) => ScriptError::DisabledOpcode(None),
-        ScriptError::SignatureEncoding(sig_err) => match sig_err {
-            signature::Error::SigHashType(Some(_)) => signature::Error::SigHashType(None).into(),
-            signature::Error::SigDER(Some(_)) => signature::Error::SigDER(None).into(),
-            signature::Error::SigHighS => ScriptError::UnknownError,
-            _ => sig_err.into(),
-        },
-        ScriptError::ReadError { .. } => ScriptError::BadOpcode(None),
-        ScriptError::ScriptNumError(_) => ScriptError::UnknownError,
-        _ => err,
-    }
-}
-
 /// Convert errors that don’t exist in the C++ code into the cases that do.
 pub fn normalize_error(err: Error) -> Error {
     match err {
-        Error::Ok(serr) => Error::Ok(normalize_script_error(serr)),
+        Error::Ok(serr) => Error::Ok(zcash_script::normalize_error(serr)),
         _ => err,
     }
 }
@@ -305,7 +291,7 @@ impl<T: ZcashScript, U: ZcashScript> ZcashScript for ComparisonInterpreter<T, U>
     ) -> Result<(), Error> {
         let (cxx, rust) =
             check_verify_callback(&self.first, &self.second, script_pub_key, script_sig, flags);
-        if rust.clone().map_err(normalize_error) != cxx {
+        if rust.clone().map_err(normalize_error) != cxx.clone().map_err(normalize_error) {
             // probably want to distinguish between
             // - one succeeding when the other fails (bad), and
             // - differing error codes (maybe not bad).
@@ -322,11 +308,10 @@ impl<T: ZcashScript, U: ZcashScript> ZcashScript for ComparisonInterpreter<T, U>
 pub mod testing {
     use crate::{
         interpreter::{State, StepFn, VerificationFlags},
-        normalize_script_error,
         script::{self, Operation, Script},
         script_error::ScriptError,
         test_vectors::TestVector,
-        Error,
+        zcash_script, Error,
     };
 
     /// Ensures that flags represent a supported state. This avoids crashes in the C++ code, which
@@ -384,7 +369,8 @@ pub mod testing {
             Ok(()) => (),
             Err(actual) => {
                 if try_normalized_error
-                    && tv.result.clone().map_err(normalize_script_error) == actual
+                    && tv.result.clone().map_err(zcash_script::normalize_error)
+                        == actual.clone().map_err(zcash_script::normalize_error)
                 {
                     ()
                 } else {
@@ -464,7 +450,10 @@ mod tests {
             flags,
         );
 
-        assert_eq!(ret.0, ret.1.map_err(normalize_error));
+        assert_eq!(
+            ret.0.clone().map_err(normalize_error),
+            ret.1.map_err(normalize_error)
+        );
         assert!(ret.0.is_ok());
     }
 
@@ -494,7 +483,10 @@ mod tests {
             flags,
         );
 
-        assert_eq!(ret.0, ret.1.map_err(normalize_error));
+        assert_eq!(
+            ret.0.clone().map_err(normalize_error),
+            ret.1.map_err(normalize_error)
+        );
         assert_eq!(ret.0, Err(Error::Ok(ScriptError::EvalFalse)));
     }
 
@@ -525,7 +517,10 @@ mod tests {
             flags,
         );
 
-        assert_eq!(ret.0, ret.1.map_err(normalize_error));
+        assert_eq!(
+            ret.0.clone().map_err(normalize_error),
+            ret.1.map_err(normalize_error)
+        );
         assert_eq!(ret.0, Err(Error::Ok(ScriptError::EvalFalse)));
     }
 
@@ -592,7 +587,7 @@ mod tests {
                 &sig[..],
                 flags,
             );
-            prop_assert_eq!(ret.0, ret.1.clone().map_err(normalize_error),
+            prop_assert_eq!(ret.0.map_err(normalize_error), ret.1.clone().map_err(normalize_error),
                             "original Rust result: {:?}", ret.1);
         }
 
@@ -627,7 +622,7 @@ mod tests {
                 &sig[..],
                 flags,
             );
-            prop_assert_eq!(ret.0, ret.1.clone().map_err(normalize_error),
+            prop_assert_eq!(ret.0.map_err(normalize_error), ret.1.clone().map_err(normalize_error),
                             "original Rust result: {:?}", ret.1);
         }