+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `interpreter::eval_step`,
   - `interpreter::verify_script`, and
   - `stepwise_verify`.
+- `ScriptError::Ok` has been removed, as it was never used
 
 ### Changed
 
@@ -39,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `interpreter::Signature` is now `signature::Decoded`
 - `interpreter::SignatureChecker::check_sig` takes a `&signature::Decoded` instead of a `&[u8]`
 - `signature::HashType` now has private fields, but there are corresponding getter methods
+- `ScriptError::UnknownError` has been replaced by `ScriptError::ExternalError`
+- `BadOpcode` and `DisabledOpcode` in `ScriptError` now take parameters
 
 ### Added